+++ src/test_utils/mod.rs
@@ -3,7 +3,7 @@ use crate::{
     Error,
 };
 
-#[cfg(feature = "interface")]
+#[cfg(feature = "serialization")]
 pub mod non_regression;
 
 pub fn policy() -> Result<Policy, Error> {
@@ -220,9 +220,11 @@ mod tests {
         )?;
 
         // Decryption fails without refreshing the user key
-        assert!(encrypted_header
-            .decrypt(&cover_crypt, &top_secret_fin_usk, None)
-            .is_err());
+        assert!(
+            encrypted_header
+                .decrypt(&cover_crypt, &top_secret_fin_usk, None)
+                .is_err()
+        );
 
         cover_crypt.refresh_user_secret_key(
             &mut top_secret_fin_usk,
@@ -235,9 +237,11 @@ mod tests {
         )?;
 
         // The refreshed key can decrypt the header
-        assert!(encrypted_header
-            .decrypt(&cover_crypt, &top_secret_fin_usk, None)
-            .is_ok());
+        assert!(
+            encrypted_header
+                .decrypt(&cover_crypt, &top_secret_fin_usk, None)
+                .is_ok()
+        );
 
         Ok(())
     }